+++ src/lemmings/data/sprites.rs
@@ -29,10 +29,10 @@ macro_rules! iterate_bits { ($data:expr) => { $data.iter().flat_map(|byte| { ret
 // Sprites are stored as 4 planes, eg all the 1 bits, then all the 2 bits, then so on. It seems they did this so
 // they could reuse one of the bits as the mask plane. But it means parsing is weird.
 // Palette entries are 0xRRGGBBAA
-pub fn extract(data: &[u8], width: u8, height: u8, image_loc: u16, mask_loc: u16, palette: &[u32; 16]) -> Vec<u32> {
+pub fn extract(data: &[u8], width: usize, height: usize, image_loc: u16, mask_loc: u16, palette: &[u32; 16]) -> Vec<u32> {
     let image_data: &[u8] = &data[image_loc as usize..];
     let mask_data: &[u8] = &data[mask_loc as usize..];
-    let pixels: usize = (width as usize) * (height as usize);
+    let pixels: usize = width * height;
     let mut image_iter_0 = iterate_bits!(image_data);
     let mut image_iter_1 = iterate_bits!(image_data).skip(pixels);
     let mut image_iter_2 = iterate_bits!(image_data).skip(pixels * 2);
